initial refactor of add unit dialog into placement ui module
