+++ cloudinstall/ev.py
@@ -70,14 +70,8 @@ class EventLoop:
                 if self.config.getopt('current_state') != \
                    ControllerState.SERVICES:
                     return
-                charm_modules = utils.load_charms(
-                    self.config.getopt('charm_plugin_dir'))
-                charm_classes = [m.__charm_class__ for m in charm_modules
-                                 if m.__charm_class__.allow_multi_units and
-                                 not m.__charm_class__.disabled]
-                # FIXME: Add unecessary confusion
-                self.ui.show_add_charm_info(charm_classes,
-                                            self._callback_map['add_charm'])
+                cb = self._callback_map['add_services']
+                self.ui.show_add_services_dialog(cb)
             if key in ['q', 'Q']:
                 self.exit(0)
             if key in ['r', 'R', 'f5']:
